requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptnoxpy==2.1.1
cryptography==37.0.4
cytoolz==0.11.2
ecdsa==0.17.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.6.1
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
pycparser==2.21
pycryptodome==3.15.0
pyrsistent==0.18.1
pyscard==2.0.3
requests==2.28.1
rlp==2.0.1
RPi.GPIO==0.7.1
six==1.16.0
termcolor==1.1.0
toolz==0.11.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.2
websockets==9.1
yarl==1.7.2
